Update Candlestick Pattern Identifier
MAJOR UPDATE: The additions and removals of candlestick patterns in this update are based on research by Thomas Bulkowski in the Encyclopedia of Candlestick Patterns. I included patterns with high probability of success and removed patterns with a low probability of success. I also felt that some patterns, like the Maribozu for example, are visually easy to identify and do not need an alert for them. 

18 new patterns have been added due to their high probability of success (66% or higher) in Bull or Bear Markets: 

- Separating Lines (Bullish and Bearish)
- Three Line Strike (Bullish and Bearish) 
- Belt Hold (Bullish and Bearish) 
- Abandoned Baby (Bullish and Bearish)
- Rising and Falling Window
- Three Outside Up and Down 
- Morning and Evening Doji Star
- Above the Stomach
- Takuri Line
- Two Black Gapping
- Identical Three Black Crows 

Modifications: The logic behind all existing patterns has been modified for improved functionality and ease of readability 

- Dots have been added to improve readability by indicating the number of candles being referenced for a pattern. 
  (Ex: three dots = three candles being referenced for pattern) 

7 patterns have been removed due to ease of identification or low probability of success: 

- Bullish and Bearish Maribozu
- Tweezer Top and Bottom
- Hammer
- Hanging Man
- Shooting Star

In total, there are now 25 candlestick patterns that I believe present the most reliable signals for Bull and Bear markets.